ajustes gerais e implmentações no RED-16
