+++ web-app-vue/src/pages/configuracao/card-configuracoes.vue
@@ -151,7 +151,7 @@ import { computed, onMounted, reactive, ref } from 'vue'
 import { useRouter } from 'vue-router'
 import { useUserStore } from '@/stores/userStore'
 import { lengthRule, requiredRule } from '@/services/basic-rules'
-import { prepareByte, readTextFile } from '@/services/formaters'
+import { prepareByte, readTextFile } from '@/services/formatters'
 import { downloadCsv, uploadCsv } from '@/services/api'
 import { useContaStore } from '@/stores/contaStore'
 import { useCategoriaStore } from '@/stores/categoriaStore'
