+++ web-app-vue/src/pages/movimentacao/lista-movimentacoes.vue
@@ -184,7 +184,7 @@ import { useRouter } from 'vue-router'
 import { useMovimentacaoStore } from '@/stores/movimentacaoStore'
 import { useCategoriaStore } from '@/stores/categoriaStore'
 import { useContaStore } from '@/stores/contaStore'
-import { prepareBalance, prepareDate } from '@/services/formaters'
+import { prepareBalance, prepareDate } from '@/services/formatters'
 import { endOfMonth, format, startOfMonth } from 'date-fns'
 import CategoriaAutocomplete from '@/shared/categoria-autocomplete.vue'
 import ContaAutocomplete from '@/shared/conta-autocomplete.vue'
