+++ web-app-vue/src/pages/usuario/user-details.vue
@@ -3,27 +3,27 @@
     <v-form v-model="valid" @submit.prevent.stop="save" class="auth-form">
       <v-text-field
         :rules="[requiredRule]"
-        v-model="user.nome"
+        v-model="userUpdate.nome"
         label="Nome"
         required
       ></v-text-field>
       <v-text-field
         :rules="[requiredRule]"
-        v-model="user.email"
+        v-model="userUpdate.email"
         label="Email"
         required
         type="email"
       ></v-text-field>
       <v-text-field
-        :rules="[requiredRule]"
-        v-model="user.senha"
+        :rules="[requiredRule, minSizeRule(6)]"
+        v-model="userUpdate.senha"
         label="Senha"
         required
         type="password"
       ></v-text-field>
       <v-text-field
-        :rules="[requiredRule]"
-        v-model="user.repetirSenha"
+        :rules="[requiredRule, minSizeRule(6)]"
+        v-model="userUpdate.repetirSenha"
         label=" Confirma Senha"
         required
         type="password"
@@ -54,32 +54,41 @@
 
 import { reactive, ref } from 'vue'
 import { useUserStore } from '@/stores/userStore'
-import { requiredRule } from '@/services/basic-rules'
-import { useRouter } from 'vue-router'
+import { minSizeRule, requiredRule } from '@/services/basic-rules'
+import { useRoute, useRouter } from 'vue-router'
 
 const valid = ref(false)
 const uState = useUserStore()
 const router = useRouter()
+const route = useRoute()
 
-const user = reactive({
+const userUpdate = reactive({
+  id: uState.userData.id,
   nome: uState.userData.nome,
   email: uState.userData.email,
-  senha: '',
-  repetirSenha: ''
+  editToken: route.params.editToken,
+  repetirSenha: '',
+  senha: ''
 })
 
 const save = async () => {
   if (!valid.value) return
-  if(confirm("Deseja realmente alterar seus dados pessoais?")) {
+  if (confirm('Deseja realmente alterar seus dados pessoais?')) {
     alert('Verifique o email de confirmação de alteração de dados')
-    // TODO alterar os dados e descartar o link de alteração
-    await uState.logout()
-    await router.push('/auth')
+    try {
+      // TODO alterar os dados e descartar o link de alteração
+      await uState.update(userUpdate)
+      await uState.logout()
+      await router.push('/auth')
+    } catch (e) {
+      console.log(e)
+      alert('Problema ao atualizar')
+    }
   }
 }
 
 const cancelar = async () => {
-  if(confirm('Deseja realmente cancelar? será preciso gerar um novo link de alteração de dados!')) {
+  if (confirm('Deseja realmente cancelar? será preciso gerar um novo link de alteração de dados!')) {
     // TODO descartar o link de alteração
     await uState.logout()
     await router.push('/auth')
@@ -93,6 +102,7 @@ const cancelar = async () => {
   min-width: 300px;
   padding: 1em;
 }
+
 .row {
   display: flex;
   flex-direction: row;
