+++ web-app-vue/src/shared/chip-saldo-movimentacao.vue
@@ -8,6 +8,7 @@
     :color="saldo >= 0 ? 'green-accent-2' : 'red-accent-2'"
     @click="showing = ++showing % 3"
   >
+    {{label}} &nbsp;
     <span v-if="showing === 0">{{ prepareMoney(entrada) }}</span>
     <span v-if="showing === 1">-{{ prepareMoney(saida) }}</span>
     <span v-if="showing === 2">{{ prepareMoney(saldo) }}</span>
@@ -16,7 +17,7 @@
 <style></style>
 <script setup>
 import { computed, ref } from 'vue'
-import { prepareBalance, prepareExpense, prepareIncome, prepareMoney } from '@/services/formaters'
+import { prepareBalance, prepareExpense, prepareIncome, prepareMoney } from '@/services/formatters'
 
 const props = defineProps(['label', 'movimentacoes'])
